Move valuespec stuff
